refactor(useTimer): Timer 관련 타입 분리
