+++ client/src/hooks/use-timer.tsx
@@ -1,28 +1,6 @@
+import type { TimerConfig, TimerReturnValue } from '~types/timer';
+
 import { useState, useEffect, useCallback } from 'react';
-
-type TimerStatus = 'RUNNING' | 'PAUSED' | 'STOPPED';
-type TimerType = 'DECREMENTAL' | 'INCREMENTAL';
-
-export type TimerReturnValue = {
-  advanceTime: (timeToAdd: number) => void;
-  pause: () => void;
-  reset: () => void;
-  start: () => void;
-  status: TimerStatus;
-  time: number;
-};
-
-export interface TimerConfig {
-  autostart: boolean;
-  endTime: number;
-  initialTime: number;
-  initialStatus: TimerStatus;
-  interval: number;
-  onTimeOver?: () => void;
-  onTimeUpdate?: (time: number) => void;
-  step: number;
-  timerType: TimerType;
-}
 
 const useTimer = ({
   autostart = false,
